+++ generator/page_generators/documentation/panther/intrinsics/intrinsics.js
@@ -27,6 +27,16 @@ let page = new Page(__filename, {
 });
 
 
+page.paragraph("Instrinsics are any symbol (excluding types) that is builtin. They mainly exist add some sort of functionality into the language such as reflection or a runtime operation that otherwise would require inline assembly.");
+
+
+
+page.h2("Intrinsic Functions");
+
+page.paragraph(`Intrinsic functions are generally not meant to be used by the programmer as there is usually a better way of doing the same thing. For example, type conversion intrinsic functions do exist but it is preferred to use ${terms.get("operator as")} instead.`);
+
+page.paragraph(`Intrinsic functions are almost always inlined and some (such as type trait intrinsics) don't have any runtime execution at all. As such, you cannot take the address of an intrinsic function.`);
+
 page.text(html.link("Arithmetic", "/site/documentation/panther/intrinsics/arithmetic_intrinsics.html"));
 page.text(html.link("Bitwise", "/site/documentation/panther/intrinsics/bitwise_intrinsics.html"));
 // page.text(html.link("Build System", "/site/documentation/panther/intrinsics/build system_intrinsics.html"));